add downloading and generation functions

+++ demo/src/main/java/com/monetware/demo/xml/CreateXML.java
@@ -1,8 +1,9 @@
-package com.monetware.demo.CreateXML;
+package com.monetware.demo.xml;
 
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.OutputStream;
+import com.monetware.demo.variable.root;
 
 public class CreateXML {
 
@@ -14,19 +15,17 @@ public class CreateXML {
      * and you will create a test.xml under the same folder
      * @throws Exception
      */
-    public void CreateFile(String fileName) throws Exception{
-
-        //file root
-        String fileroot = "C:/Users/Stephen/Desktop/dataverse/demo/src/main/resources/static/files/";
-
-        File file = new File(fileroot + "tab/" + fileName + ".tab");
+    public String CreateFile(String fileName) throws Exception{
+        File file = new File(root.FileRoot + "tab/" + fileName + ".tab");
         WriteXML c = new WriteXML();
 
         //write xml context into os
         OutputStream os = c.getOutputStream(file, "1", "2", "3", "4", "5","text/tab-separated-values");
 
         //define output file
-        File fileout = new File(fileroot + "xml/" + fileName + ".xml");
+        File fileout = new File(root.FileRoot + "xml/" + fileName + "-ddi.xml");
+        if (fileout.exists())
+            return "xml file already exist!";
         FileOutputStream fos = new FileOutputStream(fileout);
 
         //convert os into byte array and store them into test.xml
@@ -35,5 +34,6 @@ public class CreateXML {
         fos.write(flush);
         fos.close();
         os.close();
+        return "create file successfully!";
     }
 }
